requirements.txt
allure-pytest==2.14.2
pytest==8.3.5
requests==2.32.3
selenium==4.26.0
python-dotenv==1.1.0
pydantic==2.11.5
faker==37.3.0